actions/requirements2.txt
python-dotenv
requests
rasa
rasa-sdk
